requirements.txt
PyJWT
werkzeug
Flask
requests
SQLAlchemy